Data Export Changes : HTTP request will not be not wait. Refractored related classes
