+++ src/main/java/org/recap/repository/ETLRequestLogDetailsRepository.java
@@ -6,8 +6,7 @@ import org.recap.repository.jpa.BaseRepository;
 import java.util.List;
 
 public interface ETLRequestLogDetailsRepository extends BaseRepository<ETLRequestLogEntity> {
-    ETLRequestLogEntity findByEtlStatusId(Integer statusId);
-    List<ETLRequestLogEntity> findAllByEtlStatusId(Integer statusId);
-
-    List<ETLRequestLogEntity> findAllByEtlStatusIdOrderByRequestedTime(Integer statusId);
+    List<ETLRequestLogEntity> findAllByExportStatusId(Integer statusId);
+    List<ETLRequestLogEntity> findByExportStatusIdAndTransmissionType(Integer statusId, String fetchTypeId);
+    List<ETLRequestLogEntity> findByExportStatusIdAndTransmissionTypeOrderByRequestedTime(Integer statusId, String fetchTypeId);
 }
