+++ src/main/java/org/recap/service/transmission/datadump/DataDumpS3TransmissionService.java
@@ -38,7 +38,7 @@ public class DataDumpS3TransmissionService implements DataDumpTransmissionInterf
      */
     @Override
     public boolean isInterested(DataDumpRequest dataDumpRequest) {
-        return dataDumpRequest.getTransmissionType().equals(RecapConstants.DATADUMP_TRANSMISSION_TYPE_FTP);
+        return dataDumpRequest.getTransmissionType().equals(RecapConstants.DATADUMP_TRANSMISSION_TYPE_S3);
     }
 
     /**
